feat: fix build err
